# .synnovator/openapi.yaml
openapi: 3.0.3
info:
  title: Synnovator API
  version: 1.0.0
  description: API for Synnovator - Activity and Competition Management Platform
  contact:
    name: API Support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
- url: http://localhost:8000
  description: Development server
- url: https://api.synnovator.com
  description: Production server
...
paths:
...
  /users:
    get:
      summary: List users
      operationId: list_users
      tags:
      - users
      parameters:
      - $ref: '#/components/parameters/SkipParam'
      - $ref: '#/components/parameters/LimitParam'
      - name: role
        in: query
        schema:
          $ref: '#/components/schemas/UserRole'
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUserList'
    post:
      summary: Create user
      operationId: create_user
      tags:
      - users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /users/{user_id}:
    get:
      summary: Get user
      operationId: get_user
      tags:
      - users
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update user
      operationId: update_user
      tags:
      - users
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete user
      operationId: delete_user
      tags:
      - users
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /users/{user_id}/follow:
    post:
      summary: Follow user
      operationId: follow_user
      tags:
      - users
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '201':
          description: Followed
        '409':
          description: Already following
    delete:
      summary: Unfollow user
      operationId: unfollow_user
      tags:
      - users
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Unfollowed
        '404':
          $ref: '#/components/responses/NotFound'
  /users/{user_id}/followers:
    get:
      summary: List user followers
      operationId: list_user_followers
      tags:
      - users
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
      - $ref: '#/components/parameters/SkipParam'
      - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: List of followers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUserList'
  /users/{user_id}/following:
    get:
      summary: List users that this user follows
      operationId: list_user_following
      tags:
      - users
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
      - $ref: '#/components/parameters/SkipParam'
      - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: List of users being followed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUserList'
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: /oauth/authorize
          tokenUrl: /oauth/token
          scopes:
            read: Read access to resources
            write: Write access to resources
            admin: Admin access for batch operations
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
              example: NOT_FOUND
            message:
              type: string
              example: Resource not found
            details:
              type: object
              additionalProperties: true
...
    UserRole:
      type: string
      enum:
      - participant
      - organizer
      - admin
      description: User role
...
    UserCreate:
      type: object
      required:
      - username
      - email
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        display_name:
          type: string
        avatar_url:
          type: string
          format: uri
        bio:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
    UserUpdate:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        display_name:
          type: string
        avatar_url:
          type: string
          format: uri
        bio:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
    User:
      type: object
      required:
      - id
      - username
      - email
      - role
      - created_at
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        display_name:
          type: string
        avatar_url:
          type: string
          format: uri
        bio:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
...
    PaginatedUserList:
      type: object
      required:
      - items
      - total
      - skip
      - limit
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/User'
        total:
          type: integer
          description: Total number of records
        skip:
          type: integer
        limit:
          type: integer
...
  parameters:
    SkipParam:
      name: skip
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
      description: Number of records to skip
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Maximum number of records to return
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
security:
- oauth2:
  - read
  - write